Implement limiting the number of shard types
Implement limiting the number of shards with a given `custom_metadata`
resulting in using at most a `custom_metadata_type_limit` shards of the
same `custom_metadata`. This is useful for experiments with how much
data is needed to train a neural network (say with `custom_metadata`
denoting the uniform content type in the whole shard).