Fix Rust benchmarking
Pull request https://github.com/google/sedpack/pull/171 introduced a
problem with benchmarking code rendering it useless. This commit
traverses the dataset directory recursively and thus again introducec
meaningful benchmarking. Also it makes sure that we traverse the
expected number of shards.

This commit includes also test and holdout since when we have that data
we might use it. This will introduce a regression compared to the first
benchmarks and definitely one compared to the empty benchmarks.